Create utility file for provider injection
